+++ 4337-gas-metering/.env.example
@@ -27,9 +27,7 @@ ALCHEMY_GAS_POLICY = "" # https://dashboard.alchemy.com/gas-manager
 ALCHEMY_ENTRYPOINT_ADDRESS = "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789"
 ALCHEMY_MULTISEND_ADDRESS = "0x38869bf66a61cF6bDB996A6aE40D5853Fd43B526"
 # Make sure all nonce are unique for it to deploy account when run initially.
-ALCHEMY_ACCOUNT_PAYMASTER_NONCE = "4"
-ALCHEMY_ERC20_PAYMASTER_NONCE = "5"
-ALCHEMY_ERC721_PAYMASTER_NONCE = "6"
+ALCHEMY_NONCE = "2"
 # Alchemy Token Operation Values (Based on the chain selected, these values should also change accordingly.)
 ALCHEMY_ERC20_TOKEN_CONTRACT = ""
 ALCHEMY_ERC721_TOKEN_CONTRACT = ""
